Fix validation on ONE_TO_ONE Relation
Bevahiour: An One To One relation form isn't validated for a unique
value, and throw an exception when Doctrine try to insert it.

Fix: The validation has been added, but another fix had to be done
handle it. When the form was validated to false, every relation field
was deleted from the form and wasn't rendered anymore. I'd to change
this behaviour by adding a DataTransformer, and remove form generation
from PRE_SET_DATA event. The form is now generated with the FormBuilder,
and an AjaxForm now implement the new DataTransformer. This change has
an impact on forms generated by Pum.